Refactored filterBuilders move some to separate file
to get a better overview
