+++ frontend/shared/components/src/filters/filterBuilders.ts
@@ -1,5 +1,5 @@
-import type { AppType, EventNotificationType, Group, LoadedPermissions, Organization, Platform, PrivateWebshop, RecordCategory, WebshopPreview } from '@stamhoofd/structures';
-import { AuditLogType, BalanceItemStatus, BalanceItemType, CheckoutMethodType, CheckoutMethodTypeHelper, DocumentStatus, DocumentStatusHelper, EventNotificationStatus, EventNotificationStatusHelper, FilterWrapperMarker, Gender, getAuditLogTypeName, getBalanceItemStatusName, getBalanceItemTypeName, OrderStatus, OrderStatusHelper, PaymentMethod, PaymentMethodHelper, PaymentStatus, PaymentStatusHelper, PaymentType, PaymentTypeHelper, RecordType, Webshop } from '@stamhoofd/structures';
+import type { AppType, EventNotificationType, Group, LoadedPermissions, Organization, Platform } from '@stamhoofd/structures';
+import { AuditLogType, BalanceItemStatus, BalanceItemType, DocumentStatus, DocumentStatusHelper, EventNotificationStatus, EventNotificationStatusHelper, FilterWrapperMarker, Gender, getAuditLogTypeName, getBalanceItemStatusName, getBalanceItemTypeName, PaymentMethod, PaymentMethodHelper, PaymentStatus, PaymentStatusHelper, PaymentType, PaymentTypeHelper } from '@stamhoofd/structures';
 import { Formatter } from '@stamhoofd/utility';
 import { computed } from 'vue';
 import { useContext, useOrganization, usePlatform } from '../hooks';
@@ -10,6 +10,7 @@ import { NumberFilterBuilder, NumberFilterFormat } from './NumberUIFilter';
 import { SimpleNumberFilterBuilder } from './SimpleNumberUIFilter';
 import { StringFilterBuilder } from './StringUIFilter';
 import type { UIFilter, UIFilterBuilder, UIFilterBuilders } from './UIFilter';
+import { getFilterBuildersForRecordCategories } from './filter-builders/record-categories';
 
 export const getCustomerUIFilterBuilders: () => UIFilterBuilders = () => {
     const builders: UIFilterBuilders = [
@@ -621,266 +622,7 @@ export function useAuditLogUIFilterBuilders() {
     return all;
 }
 
-//
-// CHECKOUT
-//
-function getCartFilterBuilder(webshop: Webshop) {
-    return new GroupUIFilterBuilder({
-        name: $t(`%1DQ`),
-        description: $t('%17R'),
-        builders: [
-            new NumberFilterBuilder({
-                name: $t(`%17T`),
-                type: NumberFilterFormat.Number,
-                key: 'amount',
-            }),
-            new MultipleChoiceFilterBuilder({
-                name: $t(`%Sc`),
-                options: webshop.products.map((product) => {
-                    return new MultipleChoiceUIFilterOption(product.name, product.id);
-                }),
-                wrapper: {
-                    product: {
-                        id: {
-                            $in: FilterWrapperMarker,
-                        },
-                    },
-                },
-            }),
-            ...webshop.products.filter(product => product.prices.length > 1).map(product => new MultipleChoiceFilterBuilder({
-                name: product.name + ' (' + $t('%17S') + ')',
-                options: product.prices.map((price) => {
-                    return new MultipleChoiceUIFilterOption(price.name, price.id);
-                }),
-                wrapper: {
-                    product: {
-                        id: product.id,
-                    },
-                    productPrice: {
-                        id: {
-                            $in: FilterWrapperMarker,
-                        },
-                    },
-                },
-            })),
-        ],
-        wrapper: {
-            items: {
-                $elemMatch: FilterWrapperMarker,
-            },
-        },
-    });
-}
-export function useCheckoutInMemoryFilterBuilders() {
-    return (webshop: Webshop) => {
-        const all: UIFilterBuilders = [
-            getCartFilterBuilder(webshop),
-        ];
 
-        if (webshop.meta.checkoutMethods.length > 1) {
-            all.push(
-                new MultipleChoiceFilterBuilder({
-                    name: $t(`%1Mf`),
-                    options: webshop.meta.checkoutMethods.map((method) => {
-                        return new MultipleChoiceUIFilterOption(method.typeName + ': ' + method.name, method.id);
-                    }),
-                    wrapper: {
-                        checkoutMethod: {
-                            id: {
-                                $in: FilterWrapperMarker,
-                            },
-                        },
-                    },
-                }),
-            );
-        }
-
-        // Also include complex filters
-        all.push(...getFilterBuildersForRecordCategories(webshop.meta.recordCategories));
-
-        // Recursive: self referencing groups
-        all.unshift(
-            new GroupUIFilterBuilder({
-                builders: all,
-            }),
-        );
-
-        return all;
-    };
-}
-
-export function getWebshopOrderUIFilterBuilders(preview: PrivateWebshop | WebshopPreview) {
-    const builders: UIFilterBuilders = [
-        new NumberFilterBuilder({
-            name: '#',
-            key: 'number',
-        }),
-        new MultipleChoiceFilterBuilder({
-            name: $t(`%1A`),
-            options: Object.values(OrderStatus)
-                .filter(s => s !== OrderStatus.Deleted)
-                .map((status) => {
-                    return new MultipleChoiceUIFilterOption(Formatter.capitalizeFirstLetter(OrderStatusHelper.getName(status)), status);
-                }),
-            wrapper: {
-                status: {
-                    $in: FilterWrapperMarker,
-                },
-            },
-        }),
-        new StringFilterBuilder({
-            name: $t(`%Gq`),
-            key: 'name',
-        }),
-        new StringFilterBuilder({
-            name: $t(`%1FK`),
-            key: 'email',
-        }),
-    ];
-
-    if (preview.meta.phoneEnabled) {
-        builders.push(new StringFilterBuilder({
-            name: $t(`%wD`),
-            key: 'phone',
-        }));
-    }
-
-    builders.push(new MultipleChoiceFilterBuilder({
-        name: $t(`%M7`),
-        options: Object.values(PaymentMethod).map((paymentMethod) => {
-            return new MultipleChoiceUIFilterOption(PaymentMethodHelper.getNameCapitalized(paymentMethod), paymentMethod);
-        }),
-        wrapper: {
-            paymentMethod: {
-                $in: FilterWrapperMarker,
-            },
-        },
-    }));
-
-    builders.push(new DateFilterBuilder({
-        key: 'paidAt',
-        name: $t('%1Jb'),
-    }));
-
-    if (preview.meta.checkoutMethods.length > 1) {
-        builders.push(new MultipleChoiceFilterBuilder({
-            name: $t(`%1Mf`),
-            options: preview.meta.checkoutMethods.map((checkoutMethod) => {
-                const name = Formatter.capitalizeFirstLetter(`${CheckoutMethodTypeHelper.getName(checkoutMethod.type)}: ${checkoutMethod.name}`);
-                return new MultipleChoiceUIFilterOption(name, checkoutMethod.id);
-            }),
-            wrapper: {
-                checkoutMethodId: { $in: FilterWrapperMarker },
-            },
-        }));
-    }
-
-    builders.push(
-        new DateFilterBuilder({
-            name: $t(`%cB`),
-            key: 'validAt',
-        }),
-        new NumberFilterBuilder({
-            name: $t(`%1JL`),
-            key: 'totalPrice',
-            type: NumberFilterFormat.Currency,
-        }),
-        new NumberFilterBuilder({
-            name: $t(`%M4`),
-            key: 'amount',
-        }),
-    );
-
-    const timeCount = Formatter.uniqueArray(preview.meta.checkoutMethods.flatMap(method => method.timeSlots.timeSlots).map(t => t.timeRangeString())).length;
-    const dateCount = Formatter.uniqueArray(preview.meta.checkoutMethods.flatMap(method => method.timeSlots.timeSlots).map(t => t.dateString())).length;
-
-    const hasDelivery = preview.meta.checkoutMethods.some(method => method.type === CheckoutMethodType.Delivery);
-
-    // Count checkoutmethods that are not delivery
-    const nonDeliveryCount = preview.meta.checkoutMethods.filter(method => method.type !== CheckoutMethodType.Delivery).length;
-
-    if (dateCount > 1) {
-        builders.push(
-            new DateFilterBuilder({
-                name: (hasDelivery && nonDeliveryCount > 0) ? $t(`%cC`) : (hasDelivery ? $t(`%cD`) : $t(`%cE`)),
-                key: 'timeSlotDate',
-            }));
-    }
-
-    if (timeCount > 1) {
-        // todo: change sort of timeSlotTime => should take start time into account => composite key or generated index maybe?
-        // todo: maybe group
-        builders.push(
-            new NumberFilterBuilder({
-                name: $t(`%cF`),
-                key: 'timeSlotEndTime',
-                type: NumberFilterFormat.TimeMinutes,
-            }));
-
-        builders.push(
-            new NumberFilterBuilder({
-                name: $t(`%cG`),
-                key: 'timeSlotStartTime',
-                type: NumberFilterFormat.TimeMinutes,
-            }));
-    }
-
-    builders.push(new StringFilterBuilder({
-        key: 'code',
-        name: $t('%1MX'),
-        wrapper: {
-            discountCodes: {
-                $elemMatch: FilterWrapperMarker,
-            },
-        },
-    }));
-
-    if (preview.hasTickets) {
-        builders.push(new MultipleChoiceFilterBuilder({
-            name: $t('%1MY'),
-            options: preview.hasSingleTickets
-                ? [
-                        new MultipleChoiceUIFilterOption($t('%1MZ'), 'none'),
-                        new MultipleChoiceUIFilterOption($t('%V1'), 'all'),
-
-                    ]
-                : [
-                        new MultipleChoiceUIFilterOption($t('%1Ma'), 'none'),
-                        new MultipleChoiceUIFilterOption($t('%1Mb'), 'partial'),
-                        new MultipleChoiceUIFilterOption($t('%1Mc'), 'all'),
-                    ],
-            wrapper: {
-                ticketScanStatus: {
-                    $in: FilterWrapperMarker,
-                },
-            },
-        }));
-
-        builders.push(new DateFilterBuilder({
-            key: 'ticketScannedAt',
-            name: $t('%1Md'),
-        }));
-
-        if (!preview.hasSingleTickets) {
-            builders.push(new NumberFilterBuilder({
-                name: $t('%1Me'),
-                key: 'ticketCount',
-
-            }));
-        }
-    }
-
-    if (preview instanceof Webshop) {
-        builders.push(getCartFilterBuilder(preview));
-    }
-
-    // Also include complex filters
-    builders.push(...getFilterBuildersForRecordCategories(preview.meta.recordCategories));
-
-    const groupFilter = new GroupUIFilterBuilder({ builders });
-
-    return [groupFilter, ...builders];
-}
 
 export function getDocumentsUIFilterBuilders() {
     const builders: UIFilterBuilders = [
@@ -914,140 +656,6 @@ export function getDocumentsUIFilterBuilders() {
     const groupFilter = new GroupUIFilterBuilder({ builders });
 
     return [groupFilter, ...builders];
-}
-
-export function getFilterBuildersForRecordCategories(categories: RecordCategory[], prefix = '', options?: { includeNullable?: boolean }) {
-    const all: UIFilterBuilder<UIFilter>[] = [];
-
-    for (const category of categories) {
-        const allForCategory: UIFilterBuilder<UIFilter>[] = [];
-        const categoryPrefix = category.name + ' → ';
-
-        for (const record of category.records) {
-            if (record.type === RecordType.Text || record.type === RecordType.Textarea) {
-                allForCategory.push(
-                    new StringFilterBuilder({
-                        name: prefix + categoryPrefix + record.name,
-                        key: 'value',
-                        wrapper: {
-                            recordAnswers: {
-                                [record.id]: FilterWrapperMarker,
-                            },
-                        },
-                    }),
-                );
-            }
-
-            if (record.type === RecordType.Integer || record.type === RecordType.Price) {
-                allForCategory.push(
-                    new NumberFilterBuilder({
-                        name: prefix + categoryPrefix + record.name,
-                        key: 'value',
-                        type: record.type === RecordType.Price ? NumberFilterFormat.Currency : NumberFilterFormat.Number,
-                        wrapper: {
-                            recordAnswers: {
-                                [record.id]: FilterWrapperMarker,
-                            },
-                        },
-                    }),
-                );
-            }
-
-            if (record.type === RecordType.Checkbox) {
-                const extra = options?.includeNullable
-                    ? [new MultipleChoiceUIFilterOption($t('%1CJ'), null)]
-                    : [];
-                allForCategory.push(
-                    new MultipleChoiceFilterBuilder({
-                        name: prefix + categoryPrefix + record.name,
-                        options: [
-                            ...extra,
-                            new MultipleChoiceUIFilterOption($t('%BM'), true),
-                            new MultipleChoiceUIFilterOption($t('%BN'), false),
-                        ],
-                        wrapper: {
-                            recordAnswers: {
-                                [record.id]: {
-                                    selected: { $in: FilterWrapperMarker },
-                                },
-                            },
-                        },
-                    }),
-                );
-            }
-
-            if (record.type === RecordType.ChooseOne) {
-                const extra = options?.includeNullable
-                    ? [new MultipleChoiceUIFilterOption($t('%1CJ'), null)]
-                    : [];
-
-                allForCategory.push(
-                    new MultipleChoiceFilterBuilder({
-                        name: prefix + categoryPrefix + record.name,
-                        options: [
-                            ...extra,
-                            ...record.choices.map(c => new MultipleChoiceUIFilterOption(c.name.toString(), c.id)),
-                        ],
-                        wrapper: {
-                            recordAnswers: {
-                                [record.id]: {
-                                    selectedChoice: {
-                                        id: {
-                                            $in: FilterWrapperMarker,
-                                        },
-                                    },
-                                },
-                            },
-                        },
-                    }),
-                );
-            }
-
-            if (record.type === RecordType.MultipleChoice) {
-                const extra = options?.includeNullable
-                    ? [new MultipleChoiceUIFilterOption($t('%1CJ'), null)]
-                    : [];
-
-                allForCategory.push(
-                    new MultipleChoiceFilterBuilder({
-                        name: prefix + categoryPrefix + record.name,
-                        options: [
-                            ...extra,
-                            ...record.choices.map(c => new MultipleChoiceUIFilterOption(c.name.toString(), c.id)),
-                        ],
-                        wrapper: {
-                            recordAnswers: {
-                                [record.id]: {
-                                    selectedChoices: {
-                                        id: {
-                                            $in: FilterWrapperMarker,
-                                        },
-                                    },
-                                },
-                            },
-                        },
-                    }),
-                );
-            }
-        }
-
-        allForCategory.push(
-            ...getFilterBuildersForRecordCategories(category.childCategories, prefix + categoryPrefix, options),
-        );
-
-        if (allForCategory.length > 0) {
-            const group = new GroupUIFilterBuilder({
-                name: category.name.toString(),
-                builders: allForCategory,
-            });
-            allForCategory.unshift(group);
-            all.push(
-                group,
-            );
-        }
-    }
-
-    return all;
 }
 
 /**
